Re-structure the Verify methods to include all functionality
One of the missing elements was viewing a single verification. To accomplish this the old `verify` method was moved to be called `verify_verify` (I know, a bit strange, but it works), while the `verify` method now only views a specific verification.

The `verify_create` method was altered to only require the required paramater from the API (`recipient`). The `originator` paramater is now an optional.